🖥️ ui(web): isolate code tips from page transitions
