+++ apps/web/src/ui/global/widgets/components/PageTransitionAnimationWithTips/code-tips.ts
